Zoom nuevo de cámara

+++ Classes/Effects/AllIn/AllInEffect.gd
@@ -14,6 +14,8 @@ func start():
 	lightTween.set_ease(Tween.EASE_OUT)
 
 func end():
+	if not is_instance_valid(portal):
+		return
 	portal.kill()
 	var lightTween := get_tree().create_tween()
 	lightTween.tween_property(arena.getMainLight(), "light_energy", 1, 1)
